Troubleshooting; renamed files to try running bin/tp_cli
