+++ lib/tp_cli.rb
@@ -1,26 +1,15 @@
 require 'typhoeus'
-require "tp_cli/version"
+require "tp-cli/version"
 
 module TpCommandLine
   class ActivityTrack
-    def initialize
-      if ARGV.empty?
-        puts "\n Please specificy action: ./typhscript [note] or ./typhscript [cwd]"
-
-      elsif ARGV[0] == "note" && ARGV[1].is_a?(String)
-        post_to_TP(ARGV[1])
-
-      else ARGV[0] == "cwd"
-        post_to_TP("Changed working directory")
-      end
-    end
 
     def post_to_TP(description)
       request = Typhoeus::Request.new(
         "http://localhost:3000/activities",
         method: :post,
         params: { activity: {
-          description: description,.
+          description: description,
           project_id: 1,
           source: "API"
           }
@@ -30,14 +19,15 @@ module TpCommandLine
       request.run
     end
 
-    # if ARGV.empty?
-    #   puts "\n Please specificy action: ./typhscript [note] or ./typhscript [cwd]"
+    if ARGV.empty?
+      puts "\n Please specificy action: ./typhscript [note] or ./typhscript [cwd]"
 
-    # elsif ARGV[0] == "note" && ARGV[1].is_a?(String)
-    #   post_to_TP(ARGV[1])
+    elsif ARGV[0] == "note" && ARGV[1].is_a?(String)
+      post_to_TP(ARGV[1])
 
-    # else ARGV[0] == "cwd"
-    #   post_to_TP("Changed working directory")
-    # end
+    else ARGV[0] == "cwd"
+      post_to_TP("Changed working directory")
+    end
+
   end
 end
